# src/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Telegram Group Admin API",
    "version": "1.0.0",
    "description": "API for retrieving Telegram group and supergroup administrator information via a local Telethon service."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
  "components": {
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    }
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Health status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "status": { "type": "string" },
                    "telegram": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/admins": {
      "get": {
        "summary": "Fetch admins for a Telegram group or supergroup",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "parameters": [
          {
            "name": "target",
            "in": "query",
            "required": true,
            "schema": { "type": "string" },
            "description": "Telegram group username, e.g. @groupusername or https://t.me/groupusername"
          }
        ],
        "responses": {
          "200": {
            "description": "Admins retrieved successfully"
          }
        }
      }
    },
    "/api/admins/bulk": {
      "post": {
        "summary": "Fetch admins for multiple Telegram targets in one request",
        "security": [
          {
            "ApiKeyAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "targets": {
                    "type": "array",
                    "items": { "type": "string" },
                    "minItems": 1
                  }
                },
                "required": ["targets"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Bulk admins results",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean" },
                    "results": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  }
}